requirements.txt
Django==4.2
django-crispy-forms==2.0
crispy-bootstrap5==0.6
openpyxl==3.1.2
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.0
Pillow>=10.0